dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
appdirs==1.4.4            # via virtualenv
attrs==19.3.0             # via pytest
cfgv==3.1.0               # via pre-commit
click==7.1.2              # via pip-tools
coverage==5.1             # via pytest-cov
distlib==0.3.1            # via virtualenv
factory-boy==2.12.0       # via -r dev-requirements.in
faker==4.1.1              # via factory-boy
filelock==3.0.12          # via virtualenv
identify==1.4.21          # via pre-commit
importlib-metadata==1.7.0  # via -c requirements.txt, pluggy, pre-commit, pytest, virtualenv
more-itertools==8.4.0     # via pytest
nodeenv==1.4.0            # via pre-commit
packaging==20.4           # via pytest
pip-tools==5.2.1          # via -r dev-requirements.in
pluggy==0.13.1            # via pytest
pre-commit==2.6.0         # via -r dev-requirements.in
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.0        # via -r dev-requirements.in
pytest==5.4.3             # via -r dev-requirements.in, pytest-cov
python-dateutil==2.8.1    # via -c requirements.txt, faker
pyyaml==5.3.1             # via pre-commit
six==1.13.0               # via -c requirements.txt, packaging, pip-tools, python-dateutil, virtualenv
text-unidecode==1.3       # via faker
toml==0.10.1              # via pre-commit
virtualenv==20.0.25       # via pre-commit
wcwidth==0.2.5            # via pytest
zipp==3.1.0               # via -c requirements.txt, importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
